requirements.txt
mutagen
PyYAML
tabulate
